--- a/Sources/LeafKit/LeafCache/LeafCache.swift
+++ b/Sources/LeafKit/LeafCache/LeafCache.swift
@@ -52,6 +52,14 @@ public protocol LeafCache {
     ) -> EventLoopFuture<ResolvedDocument>
 }
 
+/// A `LeafCache` that provides certain blocking methods for non-future access to the cache
+///
+/// Adherents *MUST* be thread-safe and *SHOULD NOT* be blocking simply to avoid futures -
+/// only adhere to this protocol if using futures is needless overhead
+internal protocol BlockingLeafCache: LeafCache {
+    func load(documentName: String) -> LeafAST?
+}
+
 // MARK: - LeafCache default implementations for older adherants
 extension LeafCache {
     /// default implementation of remove to avoid breaking custom LeafCache adopters