--- a/Sources/LeafKit/LeafRenderer.swift
+++ b/Sources/LeafKit/LeafRenderer.swift
@@ -90,20 +90,14 @@ public final class LeafRenderer {
     public func render(path: String, context: [String: LeafData]) -> EventLoopFuture<ByteBuffer> {
         guard path.count > 0 else { return self.eventLoop.makeFailedFuture(LeafError(.noTemplateExists("(no key provided)"))) }
 
-        // If cache provides blocking load, try to get a flat AST immediately
-        if let blockingCache = cache as? BlockingLeafCache,
-           let cached = blockingCache.load(documentName: path),
-           cached.flat {
-            do {
-                let buffer = try serialize(cached, context: context)
-                return eventLoop.makeSucceededFuture(buffer)
-            } catch {
-                return eventLoop.makeFailedFuture(LeafError(.unknownError("Serializer Error")))
-            }
+        // If a flat AST is cached and available, serialize and return
+        if let flatAST = getFlatCachedHit(path),
+           let buffer = try? serialize(flatAST, context: context) {
+            return eventLoop.makeSucceededFuture(buffer)
         }
         
         // Otherwise operate using normal future-based full resolving behavior
-        return self.cache.load(documentName: path, on: self.eventLoop).flatMapThrowing { cached in
+        return self.cache.retrieve(documentName: path, on: self.eventLoop).flatMapThrowing { cached in
             guard let cached = cached else { throw LeafError(.noValueForKey(path)) }
             guard cached.flat else { throw LeafError(.unresolvedAST(path, Array(cached.unresolvedRefs))) }
             return try self.serialize(cached, context: context)
@@ -122,7 +116,13 @@ public final class LeafRenderer {
     internal func render(source: String, path: String, context: [String: LeafData]) -> EventLoopFuture<ByteBuffer> {
         guard path.count > 0 else { return self.eventLoop.makeFailedFuture(LeafError(.noTemplateExists("(no key provided)"))) }
 
-        return self.cache.load(documentName: sourcePath(source, path), on: self.eventLoop).flatMapThrowing { cached in
+        // If a flat AST is cached and available, serialize and return
+        if let flatAST = getFlatCachedHit(sourcePath(source, path)),
+           let buffer = try? serialize(flatAST, context: context) {
+            return eventLoop.makeSucceededFuture(buffer)
+        }
+        
+        return self.cache.retrieve(documentName: sourcePath(source, path), on: self.eventLoop).flatMapThrowing { cached in
             guard let cached = cached else { throw LeafError(.noValueForKey(path)) }
             guard cached.flat else { throw LeafError(.unresolvedAST(path, Array(cached.unresolvedRefs))) }
             return try self.serialize(cached, context: context)
@@ -161,7 +161,7 @@ public final class LeafRenderer {
     /// Recursive calls to `fetch()` from `resolve()` must provide the chain of extended
     /// templates to prevent cyclical errors
     private func fetch(source: String? = nil, template: String, chain: [String] = []) -> EventLoopFuture<LeafAST?> {
-        return cache.load(documentName: template, on: eventLoop).flatMap { cached in
+        return cache.retrieve(documentName: template, on: eventLoop).flatMap { cached in
             guard let cached = cached else {
                 return self.read(source: source, name: template, escape: true).flatMap { ast in
                     guard let ast = ast else { return self.eventLoop.makeSucceededFuture(nil) }
@@ -244,6 +244,14 @@ public final class LeafRenderer {
             let ast = try parser.parse()
             return LeafAST(name: name, ast: ast)
         }
+    }
+    
+    private func getFlatCachedHit(source: String = "", _ path: String) -> LeafAST? {
+        // If cache provides blocking load, try to get a flat AST immediately
+        guard let blockingCache = cache as? SynchronousLeafCache,
+           let cached = try? blockingCache.retrieve(documentName: sourcePath(source, path)),
+           cached.flat else { return nil }
+        return cached
     }
     
     private func sourcePath(_ source: String, _ path: String) -> String {