@tanner0101 2020-06-25
These changes are now available in [1.0.0-rc.1.13](https://github.com/vapor/leaf-kit/releases/tag/1.0.0-rc.1.13)